fix: remove all circular dependencies

+++ shelf/src/main/java/com/rszumlas/shelf/ShelfRepository.java
@@ -12,11 +12,11 @@ import java.util.Optional;
 public interface ShelfRepository extends JpaRepository<Shelf, Long> {
 
     @Query(value = "SELECT * FROM shelf s WHERE s.vodka_id = :vodka_id")
-    Optional<Shelf> findByVodkaId(@Param("vodka_id") Long vodka_id);
+    Optional<Shelf> findByVodkaId(@Param("vodka_id") Integer vodka_id);
 
     @Modifying
     @Query(value = "UPDATE shelf s SET s.crates = :crates + s.crates WHERE s.vodka_id = :vodka_id")
     void updateCratesAmount(@Param("crates") Integer crates,
-                            @Param("vodka_id") Long vodka_id);
+                            @Param("vodka_id") Integer vodka_id);
 
 }
